Plot Nice Lookng Heatmap Using Complex Heatmap

Plot Nice Lookng Heatmap Using Complex Heatmap

ArchRHeatmap(
  mat = NULL,
  scale = FALSE,
  limits = c(min(mat), max(mat)),
  colData = NULL,
  color = paletteContinuous(set = "solarExtra", n = 100),
  clusterCols = TRUE,
  clusterRows = FALSE,
  labelCols = FALSE,
  labelRows = FALSE,
  colorMap = NULL,
  useRaster = TRUE,
  rasterQuality = 5,
  split = NULL,
  fontSizeRows = 10,
  fontSizeCols = 10,
  fontSizeLabels = 8,
  colAnnoPerRow = 4,
  showRowDendrogram = FALSE,
  showColDendrogram = FALSE,
  customRowLabel = NULL,
  customRowLabelIDs = NULL,
  customColLabel = NULL,
  customColLabelIDs = NULL,
  customLabelWidth = 0.75,
  rasterDevice = "png",
  padding = 45,
  borderColor = NA,
  draw = TRUE,
  name = "Heatmap"
)

Arguments

mat

A matrix to plot heatmap from.

scale

A boolean whether to convert to row Z-scores

limits

A vector of two values describing min and mix limits to plot

colData

A DataFrame matching the columns of the input matrix to overlay on the columns of the heatmap

color

A palette to use for heatmap continuous color scheme. See paletteContinuous.

clusterCols

A boolean describing whether to cluster columns for heatmap.

clusterRows

A boolean describing whether to cluster rows for heatmap.

colorMap

A list of color mappings matching the column names in colData.

useRaster

A boolean whether to use rastering when plotting heatmap.

rasterQuality

Raster resolution of raster. Default is set to 5 higher numbers increase resolution.

split

A vector of groupings that will split the rows of heatmap (must be equal to nrow(mat)).

fontSizeRows

A numeric value representing the font size for rownames.

fontSizeCols

A numeric value representing the font size for colnames.

fontSizeLabels

A numeric value representing the font size for labels.

colAnnoPerRow

An integer value describing the number of column annotations per row in the legend.

showRowDendrogram

A boolean whether to show row dendrogram in heatmap.

showColDendrogram

A boolean whether to show column dendrogram in heatmap.

customRowLabel

A vector of indices to custom label from rows.

customRowLabelIDs

A vector of custom labels to overlay. Should match length of customRowLabel.

customColLabel

A vector of indices to custom label from columns.

customColLabelIDs

A vector of custom labels to overlay. Should match length of customColLabel.

customLabelWidth

A numeric describing the width of the column labels.

rasterDevice

Which device to use for rastering see ComplexHeatmap.

padding

A numeric (in cm) to pad the heatmap ie adding white space so that the final plot doesnt cutoff.

borderColor

A character representing the border color for each cell in the heatmap.

draw

A boolean whether to draw the heatmap immediately. If FALSE this will return a ComplexHeatmap object.

name

A character that will appear above color bar legend in heatmap.