Plot ArchR Region Track

This function will plot the coverage at an input region

ArchRRegionTrack(ArchRProj, region = NULL, groupBy = "Clusters",
  useGroups = NULL, useCoverages = FALSE,
  plotSummary = c("bulkTrack", "featureTrack", "geneTrack"),
  sizes = c(10, 0.5, 4), features = NULL, geneSymbol = NULL,
  upstream = 50000, downstream = 50000, tileSize = 100,
  normMethod = "ReadsInTSS", threads = 1, ylim = NULL,
  baseSize = 7, borderWidth = 0.4, tickWidth = 0.4,
  facetbaseSize = 7, geneAnno = getGeneAnnotation(ArchRProj),
  title = "", ...)

Arguments

ArchRProj
ArchRProject

region
GRanges region that will be plotted in (if more that one first will be selected)

groupBy
use groupings for bulk/scTrack

useGroups
select a subset of groups for plotting

useCoverages
use group coverages for track plotting

plotSummary
summary of region track to be plotted

sizes
sizes corresponding to plotSummary

features
GRanges features to be plotted (ie getPeakSet(ArchRProj))

geneSymbol
if region is null plotting can be centered at gene start site corresponding to the gene symbol

upstream
bp upstream of geneStart to extend

downstream
bp downstream of geneStart to extend

tileSize
with of tiles to plot bulk/scTrack

normMethod
normMethod normalization column in cellColData to normalize bulkTrack

threads
number of threads for parallel execution

ylim
y-limits for bulkTrack

baseSize
size of font in plot

borderWidth
border width in plot

tickWidth
axis tick width in plot

geneAnno
geneAnnotation for geneTrack

title
verbose sections

...
additional args