Identify Clusters for Single Cell Data

This function will identify clusters for single cell reduced dimensions supplied or from and ArchRProject

IdentifyClusters(input, reducedDims = "IterativeLSI",
  name = "Clusters", sampleCells = NULL, seed = 1,
  method = "seurat", dimsToUse = NULL, knnAssign = 10,
  nOutlier = 20, verbose = TRUE, tstart = NULL, ...)

Arguments

input
ArchRProject or matrix for cluster identification

reducedDims
reducedDims of ArchRProject if provided

name
name of cluster column if input is ArchRProject

seed
seed

method
supported methods are Seurat and LouvainJaccard

dimsToUse
reduced dims to use

knnAssign
number of nearest neighbors for assignment of outliers and estimation

nOutlier
number of cells required for a cluster to be called if not then these will be considered an outlier

...
arguments to provide Seurat::FindClusters or ArchR:::.clustLouvain (knn = 50, jaccard = TRUE)