Add Peak Co-Accessibility to an ArchRProject

This function will add co-accessibility scores to peaks in a given ArchRProject

addCoAccessibility(
  ArchRProj = NULL,
  reducedDims = "IterativeLSI",
  dimsToUse = 1:30,
  scaleDims = NULL,
  corCutOff = 0.75,
  k = 200,
  knnIteration = 500,
  overlapCutoff = 0.8,
  maxDist = 1e+05,
  scaleTo = 10^4,
  log2Norm = TRUE,
  seed = 1,
  knnMethod = NULL,
  threads = getArchRThreads()
)

Arguments

ArchRProj
An ArchRProject object.

reducedDims
The name of the reducedDims object (i.e. "IterativeLSI") to retrieve from the designated ArchRProject.

dimsToUse
A vector containing the dimensions from the reducedDims object to use in clustering.

scaleDims
A boolean describing whether to rescale the total variance for each principal component. This is useful for minimizing the contribution of strong biases (dominating early PCs) and lowly abundant populations. However, this may lead to stronger sample-specific biases since it is over-weighting latent PCs. If NULL this will scale the dimensions depending on if this were set true when the reducedDims were created by addIterativeLSI.

corCutOff
A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to sequencing depth that is greater than the corCutOff, it will be excluded from analysis.

k
The number of k-nearest neighbors to use for creating single-cell groups for correlation analyses.

knnIteration
The number of k-nearest neighbor groupings to test for passing the supplied overlapCutoff.

overlapCutoff
The maximum allowable overlap between the current group and all previous groups to permit the current group be added to the group list during k-nearest neighbor calculations.

maxDist
The maximum allowable distance in basepairs between two peaks to consider for co-accessibility.

scaleTo
The total insertion counts from the designated group of single cells is summed across all relevant peak regions from the peakSet of the ArchRProject and normalized to the total depth provided by scaleTo.

log2Norm
A boolean value indicating whether to log2 transform the single-cell groups prior to computing co-accessibility correlations.

seed
A number to be used as the seed for random number generation required in cluster determination. It is recommended to keep track of the seed used so that you can reproduce results downstream.

knnMethod
The method to be used for k-nearest neighbor computations. Options are "nabor", "RANN", and "FNN" and the corresponding package is required.

threads
The number of threads to be used for parallel computing.