Add embedding of a reduced dimensions object in an ArchRProject

addEmbedding(ArchRProj = NULL, reducedDims = "IterativeLSI",
  embedding = "UMAP", embeddingOut = NULL, dimsToUse = NULL,
  corCutOff = 0.75, saveModel = TRUE, seed = 1, force = FALSE,
  threads = floor(detectCores()/2), embeddingParams = list(), ...)

Arguments

ArchRProj
An ArchRProject object.

reducedDims
QQQ The name of the reducedDims object to use. Possible options include "IterativeLSI", QQQ.

embedding
QQQ The name of the embedding to create. Possible options include "UMAP", "TUMAP", "RTSNE", and "FFRTSNE".

...
additional args

colorBy
QQQ colorBy cellColData or Arrays in Arrows (ie GeneScoreMatrix)

name
QQQ name of column in cellColData or Feature in Array in Arrows

log2Norm
A boolean value that indicates whether log2 Normalization should be performed on the features if they are continuous.

pal
The name or numeric index of a custom palette from ArchR_palettes to use for plotting the individual points of the embedding visualization.

size
The numeric size of points to plot.

rastr
A boolean valut that indicates that the plot should be rasterized. This does not rasterize lines and labels, just the internal portions of the plot.

quantCut
If this is not null, a quantile cut is performed to threshold the top and bottom of the distribution. This prevents skewed color scales caused by strong outliers. The format of this should be c(x,y) where x is the upper threshold and y is the lower threshold. For example, quantileCut = c(0.975,0.025) will take the top and bottom 2.5

quantHexQQQ quantile evaluation for each hex in geom_hex

discreteSetQQQ The name or numeric index of a discrete palette from ArchR_palettesdiscrete to use for plotting QQQ.

continuousSetQQQcontinuous palette for visualizing embedding

randomizeA boolean value that determines whether to randomly order the plotting of points to avoid (for ex.) all points from a single cluster being plotted as the top-most layer of the plot.

keepAxisQQQ keep x and y axis for plot

baseSizeQQQ The numeric font size to be used in the plot. This applies to all plot labels.

plotContinuousQQQ how to plot continuous features (points and hex)

plotParamsQQQ additional params to pass to ggPoint/ggHex

plotWidthQQQ plot width used for creating a consistent plot independent of legend size

plotHeightQQQ plot height used for creating a consistent plot independent of legend size