Add Gene Expression Matrix to ArrowFiles or an ArchRProject

This function, for each sample, will add gene expression values from a paired scATAC-seq + scRNA-seq multi modal assay to the ArrowFiles or ArchRProject.

addGeneExpressionMatrix(
  input = NULL,
  seRNA = NULL,
  chromSizes = getChromSizes(input),
  excludeChr = c("chrM", "chrY"),
  scaleTo = 10000,
  verbose = TRUE,
  threads = getArchRThreads(),
  parallelParam = NULL,
  force = TRUE,
  logFile = createLogFile("addGeneExpressionMatrix")
)

Arguments

input

An ArchRProject object or character vector of ArrowFiles.

seRNA

A a scRNA-seq SummarizedExperiment (cell x gene) to be integrated with the scATAC-seq data. Cell names from this object much match those of the cell names in the ArrowFiles/ArchRProject. We will add support shortly for Seurat Objects (see Seurat::as.SingleCellExperiment). The provided values MUST be in counts (integer), not log transformed.

chromSizes

A GRanges object of the chromosome lengths. See getChromSizes for more info.

excludeChr

A character vector containing the seqnames of the chromosomes that should be excluded from this analysis.

scaleTo

Each column in the calculated gene score matrix will be normalized to a column sum designated by scaleTo.

verbose

A boolean describing whether to print to console messages of progress.

threads

The number of threads to be used for parallel computing.

parallelParam

A list of parameters to be passed for biocparallel/batchtools parallel computing.

force

A boolean value indicating whether to force the matrix indicated by matrixName to be overwritten if it already exist in the given input.

logFile

The path to a file to be used for logging ArchR output.