Add TileMatrix to Arrows/ArchRProject

This function for each sample will independently compute counts for each tile per cell in the Arrow File

addImputeWeights(ArchRProj = NULL, reducedDims = "IterativeLSI",
  dimsToUse = NULL, td = 3, ka = 4, sampleCells = max(5000,
  floor(nCells(ArchRProj)/10)), k = 15, epsilon = 1, weighted = TRUE)

Arguments

input
ArchRProject or ArrowFiles

chromSizes
chromomosome sizes used for identifying number of tiles to count

windowSize
size for each window to break up each chromosome

binarize
save as a Sparse.Binary.Matrix or Sparse.Integer.Matrix

excludeChr
exclude chromosomes from this analysis

threads
number of threads

parallelParam
parallel parameters for batch style execution

force
force overwriting previous TileMatrix in ArrowFile