Create a genome annotation object for ArchR

This function will create a genome annotation object that can be used for creating ArrowFiles or an ArchRProject, etc.

createGenomeAnnotation(
  genome = NULL,
  chromSizes = NULL,
  blacklist = NULL,
  filter = TRUE,
  filterChr = c("chrM")
)

Arguments

genome

Either (i) a string that is a valid BSgenome or (ii) a BSgenome object (ie "hg38" or "BSgenome.Hsapiens.UCSC.hg38").

chromSizes

A GRanges object containing chromosome start and end coordinates.

blacklist

A GRanges object containing regions that should be excluded from analyses due to unwanted biases.

filter

A boolean value indicating whether non-standard chromosome scaffolds should be excluded. These "non-standard" chromosomes are defined by filterChrGR() and by manual annotation using the filterChr parameter.

filterChr

A character vector indicating the seqlevels that should be removed if manual removal is desired for certain seqlevels. If no manual removal is desired, filterChr should be set to NULL. If filter is set to TRUE but filterChr is set to NULL, non-standard chromosomes will still be removed as defined in filterChrGR().

Examples

if (!require("BSgenome.Hsapiens.UCSC.hg19", quietly = TRUE)) BiocManager::install("BSgenome.Hsapiens.UCSC.hg19")
#> 
#> Attaching package: ‘Biostrings’
#> The following object is masked from ‘package:base’:
#> 
#>     strsplit
library(BSgenome.Hsapiens.UCSC.hg19)

# Get Genome
genome <- BSgenome.Hsapiens.UCSC.hg19

# Create Genome Annotation
genomeAnno <- createGenomeAnnotation(genome)
#> Error in createGenomeAnnotation(genome): could not find function "createGenomeAnnotation"

# Also can create from a string if BSgenome exists
genomeAnno <- createGenomeAnnotation("hg19")
#> Error in createGenomeAnnotation("hg19"): could not find function "createGenomeAnnotation"