Calculate footprints from an ArchRProject

This function will get footprints for all samples in a given ArchRProject and return a summarized experiment object that can be used for downstream analyses

getFootprints(
  ArchRProj = NULL,
  positions = NULL,
  plotName = "Plot-Footprints",
  groupBy = "Clusters",
  useGroups = NULL,
  flank = 250,
  minCells = 25,
  nTop = NULL,
  threads = getArchRThreads(),
  verbose = TRUE,
  logFile = createLogFile("getFootprints")
)

Arguments

ArchRProj

An ArchRProject object.

positions

A list or GRangesList of GRanges containing the positions to incorporate into the footprint. Each position should be stranded.

plotName

The prefix to add to the file name for the output PDF file containing the footprint plots.

groupBy

The name of the column in cellColData used in the addGroupCoverages() function for grouping multiple cells together.

useGroups

A character vector that is used to select a subset of groups by name from the designated groupBy column in cellColData. This limits the groups used to perform footprinting.

flank

The number of basepairs from the position center (+/-) to consider as the flank.

minCells

The minimum number of cells required in a given cell group to permit footprint generation.

nTop

The number of genomic regions to consider. Only the top nTop genomic regions based on the "score" column in the GRanges object will be considered for the footprint.

threads

The number of threads to be used for parallel computing.

verbose

A boolean value that determines whether standard output includes verbose sections.

logFile

The path to a file to be used for logging ArchR output.

Examples

# Get Test ArchR Project
proj <- getTestProject()
#> Error in getTestProject(): could not find function "getTestProject"

# Get Motif Positions
positions <- getPositions(proj)
#> Error in getPositions(proj): could not find function "getPositions"

# Get Footprints
seFoot <- getFootprints(ArchRProj = proj, positions = positions, groupBy = "Clusters", minCells = 10)
#> Error in getFootprints(ArchRProj = proj, positions = positions, groupBy = "Clusters",     minCells = 10): could not find function "getFootprints"