A ggplot-based Hexplot wrapper function summary of points in a standardized manner

ggHex(x = NULL, y = NULL, color = NULL, pal = paletteContinuous(set
  = "solar_extra"), bins = 200, xlim = NULL, ylim = NULL,
  extend = 0.05, xlabel = "x", ylabel = "y", title = "",
  colorTitle = "values", baseSize = 6, ratioYX = 1, FUN = "mean",
  quantCut = c(0.01, 0.99), addPoints = FALSE, ...)

Arguments

x
A numeric vector containing the x-axis values for each point.

y
A numeric vector containing the y-axis values for each point.

color
A numeric/categorical vector containing coloring information for each point.

pal
A custom palette for continuous coloring.

bins
The number of bins to be used for plotting the hexplot. bins indicates the total number of hexagons that will fit within the surface area of the plot.

xlim
A vector of two numeric values indicating the lower and upper bounds of the x-axis on the plot.

ylim
A vector of two numeric values indicating the lower and upper bounds of the y-axis on the plot.

extend
A numeric value indicating the fraction to extend the x-axis and y-axis beyond the maximum and minimum values if xlim and ylim are not provided. For example, 0.05 will extend the x-axis and y-axis by 5

xlabelThe label to plot for the x-axis.

ylabelThe label to plot for the y-axis.

titleThe title of the plot.

colorTitleThe label to use for the legend corresponding to color.

baseSizeThe base font size to use in the plot.

ratioYXThe aspect ratio of the x and y axes on the plot.

FUNThe function to use for summarizing data into hexagons. Typically "mean" or something similar.

quantCutIf this is not null, a quantile cut is performed to threshold the top and bottom of the distribution. This prevents skewed color scales caused by strong outliers. The format of this should be c(x,y) where x is the upper threshold and y is the lower threshold. For example, quantileCut = c(0.975,0.025) will take the top and bottom 2.5addPointsA boolean value indicating whether individual points should be shown on the hexplot....additional params to pass This function will plot x,y coordinates values summarized in hexagons in a standardized manner