A ggplot-based violin plot wrapper function

A ggplot-based violin plot wrapper function

ggViolin(x = NULL, y = NULL, xlabel = NULL, ylabel = NULL,
  xOrder = NULL, addPoints = FALSE, size = 1, baseSize = 6,
  ratioYX = NULL, sampleRatio = 0.1, title = "",
  pal = paletteDiscrete(values = x, set = "stallion"), ...)

Arguments

x
A character vector containing the categorical x-axis values for each y-axis value.

y
A numeric vector containing the y-axis values for each point.

xlabel
The label to plot for the x-axis.

ylabel
The label to plot for the y-axis.

xOrder
A character vector indicating a custom order for plotting x-axis categorical values for plotting. Should contain all possible values of x in the desired order.

addPoints
A boolean value indicating whether individual points should be added to the plot using geom_quasirandom.

size
The line width for boxplot lines.

baseSize
The base font size to use in the plot.

ratioYX
The aspect ratio of the x and y axes on the plot.

sampleRatio
sampling ratio for number of dots to be shown from original data over violins to prevent over-crowding. Default is set to 0.1.

title
The title of the plot.

pal
A custom palette for discrete coloring.