Plot a Heatmap of Identified Marker Features

This function will plot a heatmap of the results from markerFeatures

markerHeatmap(
  seMarker = NULL,
  cutOff = "FDR <= 0.01 & Log2FC >= 0.5",
  log2Norm = TRUE,
  scaleTo = 10^4,
  scaleRows = TRUE,
  plotLog2FC = FALSE,
  limits = c(-2, 2),
  grepExclude = NULL,
  pal = NULL,
  binaryClusterRows = TRUE,
  clusterCols = TRUE,
  labelMarkers = NULL,
  nLabel = NULL,
  nPrint = 20,
  labelRows = FALSE,
  returnMat = FALSE,
  transpose = TRUE,
  invert = FALSE
)

Arguments

seMarker
A SummarizedExperiment object returned by markerFeatures().

cutOff
A valid-syntax logical statement that defines which marker features from seMarker will be plotted in the heatmap. cutoff can contain any of the assayNames from seMarker.

log2Norm
A boolean value indicating whether a log2 transformation should be performed on the values in seMarker prior to plotting. Should be set to TRUE for counts-based assays (but not assays like "DeviationsMatrix").

scaleTo
Each column in the assay Mean from seMarker will be normalized to a column sum designated by scaleTo prior to log2 normalization. If log2Norm is FALSE this option has no effect.

scaleRows
A boolean value that indicates whether the heatmap should display row-wise z-scores instead of raw values.

limits
A numeric vector of two numbers that represent the lower and upper limits of the heatmap color scheme.

grepExclude
A character vector or string that indicates the rownames or a specific pattern that identifies rownames from seMarker to be excluded from the heatmap.

pal
A custom continuous palette from ArchRPalettes (see paletteContinuous()) used to override the default continuous palette for the heatmap.

binaryClusterRows
A boolean value that indicates whether a binary sorting algorithm should be used for fast clustering of heatmap rows.

labelMarkers
A character vector listing the rownames of seMarker that should be labeled on the side of the heatmap.

nLabel
An integer value that indicates whether the top n features for each column in seMarker should be labeled on the side of the heatmap.

labelRows
A boolean value that indicates whether all rows should be labeled on the side of the heatmap.

returnMat
A boolean value that indicates whether the final heatmap matrix should be returned in lieu of plotting the actual heatmap.

invert
JJJ A boolean value that indicates whether the heatmap will be sorted features by features that are Log2FC < 0 and the color palette is inverted for visualization. For example, this is useful when looking for down-regulated markers (Log2FC < 0) instead of up-regulated markers (Log2FC > 0).