Continuous Color Palette

Continuous Color Palette

paletteContinuous(set = "solarExtra", n = 256, reverse = FALSE)

Arguments

set

The name of a color palette provided in the ArchRPalettes list object.

n

The number of unique colors to generate as part of this continuous color palette.

reverse

A boolean variable that indicates whether to return the palette colors in reverse order.

Examples

# Color Palette
pal <- paletteContinuous()