Plot a Heatmap of Peak Annotation Hypergeometric Enrichment in Marker Peaks.

This function will plot a heatmap of hypergeometric enrichment of a given peakAnnotation within the defined marker peaks.

plotEnrichHeatmap(
  seEnrich = NULL,
  pal = paletteContinuous(set = "comet", n = 100),
  n = 10,
  cutOff = 20,
  pMax = Inf,
  clusterCols = TRUE,
  binaryClusterRows = TRUE,
  labelRows = TRUE,
  rastr = TRUE,
  transpose = FALSE,
  returnMatrix = FALSE,
  logFile = createLogFile("plotEnrichHeatmap")
)

Arguments

seEnrich

A SummarizedExperiment object containing peak enrichment information returned by peakAnnoEnrichment().

pal

A custom continuous palette (see paletteContinuous()) used to override the default continuous palette for the heatmap.

n

The number of top enriched peakAnnotations per column from the seMarker to display in the heatmap. This number can be lowered to improve visibility of the heatmap.

cutOff

A numeric cutOff that indicates the minimum P-adj enrichment to be included in the heatmap.

pMax

A numeric representing the maximum P-adj for plotting in the heatmap.

clusterCols

A boolean indicating whether or not to cluster columns in the heatmap.

binaryClusterRows

A boolean indicating whether or not to cluster rows using binary classification in the heatmap.

labelRows

A boolean indicating whether or not to label all rows in the heatmap.

rastr

A boolean value that indicates whether the plot should be rasterized using ComplexHeatmap. This does not rasterize lines and labels, just the internal portions of the plot.

transpose

A boolean determining whether to transpose the heatmap in the plot.

returnMatrix

A boolean determining whether to return the matrix corresponding to the heatmap rather than generate a plot.

logFile

The path to a file to be used for logging ArchR output.

Examples

# Get Test ArchR Project
proj <- getTestProject()
#> Error in getTestProject(): could not find function "getTestProject"

# Get Markers
seMarker <- getMarkerFeatures(
  ArchRProj = proj, 
  useMatrix = "PeakMatrix", 
  testMethod = "binomial", 
  binarize = TRUE
)
#> Error in getMarkerFeatures(ArchRProj = proj, useMatrix = "PeakMatrix",     testMethod = "binomial", binarize = TRUE): could not find function "getMarkerFeatures"

# Get Peak Annotation Enrichment
annoEnrich <- peakAnnoEnrichment(
  seMarker = seMarker, 
  ArchRProj = proj,
  cutOff = "FDR <= 0.1 & Log2FC >= 0"
)
#> Error in peakAnnoEnrichment(seMarker = seMarker, ArchRProj = proj, cutOff = "FDR <= 0.1 & Log2FC >= 0"): could not find function "peakAnnoEnrichment"

# Multiply by 50 since this is a super small test sample
assay(annoEnrich) <- assay(annoEnrich) * 50
#> Error in assay(annoEnrich): could not find function "assay"

#Plot
p <- plotEnrichHeatmap(annoEnrich)
#> Error in plotEnrichHeatmap(annoEnrich): could not find function "plotEnrichHeatmap"

#PDF
plotPDF(p, name = "PeakAnnoEnrich", ArchRProj = proj)
#> Error in plotPDF(p, name = "PeakAnnoEnrich", ArchRProj = proj): could not find function "plotPDF"