Plot Footprints

This function will get footprints for all samples in a given ArchRProject or a properly-formatted Summarized Experiment

plotFootprints(
  seFoot = NULL,
  names = NULL,
  pal = NULL,
  flank = 250,
  flankNorm = 50,
  normMethod = "Subtract",
  smoothWindow = NULL,
  baseSize = 6,
  plot = TRUE,
  ArchRProj = NULL,
  plotName = paste0("Plot-Footprints-", normMethod),
  height = 6,
  width = 4,
  addDOC = TRUE,
  force = FALSE,
  logFile = createLogFile("plotFootprints")
)

Arguments

seFoot

A summarized experiment object containing information on footprints returned by the getFootprints() function.

names

A character vector containing the names of the transcription factors to be plotted. These should match colnames of seFoot.

pal

The name of a custom palette from ArchRPalettes to use for plotting the lines corresponding to the footprints.

flank

The number of basepairs from the position center (+/-) to consider as the flank.

flankNorm

The number of basepairs to consider at the edge of the flank region (+/-) to be used for footprint normalization.

normMethod

The name of the normalization method to use to normalize the footprint relative to the Tn5 insertion bias. Options include "none", "subtract", "divide". "Subtract" means subtracting the normalized Tn5 Bias. "Divide" means dividing the normalized Tn5 Bias.

smoothWindow

The size in basepairs of the sliding window to be used for smoothing of the footprint signal.

baseSize

A numeric specifying the baseSize of font in the plots.

plot

A boolean value indicating whether or not the footprints should be plotted (TRUE) or returned as grob objects (FALSE).

ArchRProj

An ArchRProject object to be used for plotting directory in getOutputDirectory. If no ArchRProj is supplied, then plots will be stored in a directory called "Plots" in the current working directory.

plotName

A string indicating the name/prefix of the file to be used for output plots.

height

The height in inches to be used for the output PDF file.

width

The width in inches to be used for the output PDF file.

addDOC

A boolean variable that determines whether to add the date of creation to end of the PDF file name. This is useful for preventing overwritting of old plots.

force

If many footprints are requested when plot = FALSE, please set force = TRUE. This prevents large amount of footprint plots stored as an object.

logFile

The path to a file to be used for logging ArchR output.

Examples

# Get Test ArchR Project
proj <- getTestProject()

# Get Motif Positions
positions <- getPositions(proj)

# Get Footprints
seFoot <- getFootprints(ArchRProj = proj, positions = positions, groupBy = "Clusters", minCells = 10)

# Plot Footprints
plotFootprints(seFoot, smoothWindow = 11)