Plot a Heatmap of Identified Marker Features

This function will plot a heatmap of the results from markerFeatures

plotMarkerHeatmap(
  seMarker = NULL,
  cutOff = "FDR <= 0.01 & Log2FC >= 0.5",
  log2Norm = TRUE,
  scaleTo = 10^4,
  scaleRows = TRUE,
  plotLog2FC = FALSE,
  limits = c(-2, 2),
  grepExclude = NULL,
  pal = NULL,
  binaryClusterRows = TRUE,
  clusterCols = TRUE,
  subsetMarkers = NULL,
  labelMarkers = NULL,
  nLabel = 15,
  nPrint = 15,
  labelRows = FALSE,
  returnMatrix = FALSE,
  transpose = FALSE,
  invert = FALSE,
  logFile = createLogFile("plotMarkerHeatmap")
)

Arguments

seMarker

A SummarizedExperiment object returned by getMarkerFeatures().

cutOff

A valid-syntax logical statement that defines which marker features from seMarker will be plotted in the heatmap. cutoff can contain any of the assayNames from seMarker.

log2Norm

A boolean value indicating whether a log2 transformation should be performed on the values in seMarker prior to plotting. Should be set to TRUE for counts-based assays (but not assays like "DeviationsMatrix"). This only applies if plotLog2FC = FALSE.

scaleTo

Each column in the assay Mean from seMarker will be normalized to a column sum designated by scaleTo prior to log2 normalization. If log2Norm = FALSE this option has no effect.

scaleRows

A boolean value that indicates whether the heatmap should display row-wise z-scores instead of raw values.

plotLog2FC

A boolean value that indicates whether the log2(fold change) values should be plotted on the heatmap or not. If TRUE, the log2(fold change) value (stored in the assay of seMarker labeled "Log2FC") is plotted. If FALSE, the mean value (stored in the assay of seMarker labeled "Mean") is plotted instead, with optional log normalization based on the log2Norm parameter.

limits

A numeric vector of two numbers that represent the lower and upper limits of the heatmap color scheme.

grepExclude

A character vector or string that indicates the rownames or a specific pattern that identifies rownames from seMarker to be excluded from the heatmap.

pal

A custom continuous palette from ArchRPalettes (see paletteContinuous()) used to override the default continuous palette for the heatmap.

binaryClusterRows

A boolean value that indicates whether a binary sorting algorithm should be used for fast clustering of heatmap rows.

clusterCols

A boolean value that indicates whether the columns of the marker heatmap should be clustered.

subsetMarkers

A vector of rownames from seMarker to use for subsetting of seMarker to only plot specific features on the heatmap. Note that these rownames are expected to be integers that come from rownames(rowData(seMarker)). If this parameter is used for subsetting, then the values provided to cutOff are effectively ignored.

labelMarkers

A character vector listing the rownames of seMarker that should be labeled on the side of the heatmap.

nLabel

An integer value that indicates how many of the top n features for each column in seMarker should be labeled on the side of the heatmap. To remove all feature labels, set nLabel = 0.

nPrint

If provided seMarker is from "GeneScoreMatrix" print the top n genes for each group based on how uniquely up-regulated the gene is.

labelRows

A boolean value that indicates whether all rows should be labeled on the side of the heatmap.

returnMatrix

A boolean value that indicates whether the final heatmap matrix should be returned in lieu of plotting the actual heatmap.

transpose

A boolean value that indicates whether the heatmap should be transposed prior to plotting or returning.

invert

A boolean value that indicates whether the heatmap will display the features with the lowest log2(fold change). In this case, the heatmap will display features that are specifically lower in the given cell group compared to all other cell groups. Additionally, the color palette is inverted for visualization. This is useful when looking for down-regulated markers (log2(fold change) < 0) instead of up-regulated markers (log2(fold change) > 0).

logFile

The path to a file to be used for logging ArchR output.

Examples

#Get Test Project
proj <- getTestProject()

#Get Markers
seMarker <- getMarkerFeatures(
  ArchRProj = proj, 
  useMatrix = "PeakMatrix", 
  testMethod = "binomial", 
  binarize = TRUE
)

#Plot Markers
p <- plotMarkerHeatmap(seMarker)

#PDF
plotPDF(p, name = "Marker-Heatmap", ArchRProj = proj)