Plot Peak2Gene Heatmap from an ArchRProject

This function plots side by side heatmaps of linked ATAC and Gene regions from addPeak2GeneLinks.

plotPeak2GeneHeatmap(
  ArchRProj = NULL,
  corCutOff = 0.45,
  FDRCutOff = 1e-04,
  k = 25,
  nPlot = 25000,
  limitsATAC = c(-2, 2),
  limitsRNA = c(-2, 2),
  groupBy = "Clusters",
  palGroup = NULL,
  palATAC = paletteContinuous("solarExtra"),
  palRNA = paletteContinuous("blueYellow"),
  verbose = TRUE,
  logFile = createLogFile("plotPeak2GeneHeatmap")
)

Arguments

ArchRProj
An ArchRProject object.

corCutOff
A numeric describing the minimum numeric peak-to-gene correlation to return.

FDRCutOff
A numeric describing the maximum numeric peak-to-gene false discovery rate to return.

k
An integer describing the number of k-means clusters to group peak-to-gene links prior to plotting heatmaps.

nPlot
An integer describing the maximum number of peak-to-gene links to plot in heatmap.

limitsATAC
An integer describing the maximum number of peak-to-gene links to plot in heatmap.

limitsRNA
An integer describing the maximum number of peak-to-gene links to plot in heatmap.

groupBy
The name of the column in cellColData to use for labeling KNN groupings. The maximum group appeared in the KNN groupings is used.

palGroup
A color palette describing the colors in groupBy. For example, if groupBy = "Clusters" try paletteDiscrete(ArchRProj$Clusters) for a color palette.

palATAC
A color palette describing the colors to be used for the ATAC heatmap. For example, paletteContinuous("solarExtra").

palRNA
A color palette describing the colors to be used for the RNA heatmap. For example, paletteContinuous("blueYellow").

verbose
A boolean value that determines whether standard output should be printed.

logFile
The path to a file to be used for logging ArchR output.