Plot a TSS Enrichment Plot for Each Sample

This function will plot a TSS enrichment plot for each sample. Cells in ArchRProject are the only ones used when making this plot.

plotTSSEnrichment(
  ArchRProj = NULL,
  groupBy = "Sample",
  chromSizes = getChromSizes(ArchRProj),
  TSS = getTSS(ArchRProj),
  flank = 2000,
  norm = 100,
  smooth = 11,
  pal = NULL,
  returnDF = FALSE,
  threads = getArchRThreads(),
  logFile = createLogFile("plotTSSEnrichment")
)

Arguments

ArchRProj

An ArchRProject object.

groupBy

The name of the column in cellColData to use for grouping cells together for summarizing.

chromSizes

A GRanges object of the chromosome lengths. See getChromSizes for more info.

TSS

A GRanges object containing the locations of stranded TSS regions. The default behavior is to try to retrieve this information from the geneAnnotation stored in the ArchRProject.

flank

An integer that specifies how far in bp (+/-) to extend the TSS for plotting.

norm

An integer that specifies the number of base pairs from the ends of the flanks to be used for normalization. For example if flank=2000 and norm=100, the TSS insertions will be normalized by +/- 1900-2000 bp from the TSS.

smooth

An integer that indicates the smoothing window (in basepairs) to be applied to the TSS plot.

pal

A color palette representing the groups from groupBy in TSS plot.

returnDF

A boolean value that indicates whether to return a data.frame containing the plot information instead of plotting the TSS enrichment plot.

threads

An integer specifying the number of threads to use for calculation. By default this uses the number of threads set by addArchRThreads().

logFile

The path to a file to be used for logging ArchR output.

Examples

# Get Test ArchR Project
proj <- getTestProject()
#> Error in getTestProject(): could not find function "getTestProject"

# Plot TSS
p <- plotTSSEnrichment(proj, groupBy = "Clusters")
#> Error in plotTSSEnrichment(proj, groupBy = "Clusters"): could not find function "plotTSSEnrichment"

# PDF
plotPDF(p, name = "TSS-Enrich", ArchRProj = proj)
#> Error in plotPDF(p, name = "TSS-Enrich", ArchRProj = proj): could not find function "plotPDF"