Plot a given position weight matrix as a heatmap using ggplot2

The given matrix is plotted as a heatmap using ggplot2's geom_tile.

plot_ggheatmap(pwmMat, position_labels = NULL, savePDFfilename = NULL)

Arguments

pwmMat
Matrix (usually a PWM, but can be non-normalized/any matrix) to be represented as a heatmap.

position_labels
Labels of the positions in the sequences.

savePDFfilename
Name of the file which will be saved as PDF.

Value

A ggplot object so you can simply call print or save on it later. If savePDFfilename is given, it is also saved and the ggplot object returned.

See also