Visualize the NMF basis vectors as a heatmap

Each factor in the given features matrix is separately represented as a heatmap. Since each feeature vector is a one-hot-encoded DNA sequence, it is reshaped to have 4 rows and then visualized.

viz_basis_vectors_as_heatmap(
  featuresMatrix,
  position_labels = NA,
  add_pseudo_counts = FALSE,
  savePDFfilename = NULL,
  sinuc_or_dinuc = "sinuc"
)

Arguments

featuresMatrix
The features matrix output from NMF. Expected dimensionality: number of columns represent the number of factors from NMF, and the number of rows is 4 times the length of the sequences in the collection.

position_labels
Labels of the positions in the sequences.

add_pseudo_counts
Boolean, taking values TRUE/T or FALSE/F, default set to FALSE. Setting it to TRUE will enable adding pseudo-counts to the features matrix.

savePDFfilename
Name of the file which will be saved as PDF (also provide the extension).

sinuc_or_dinuc
"sinuc" or "dinuc"

Value

nothing.

See also