Visualize raw DNA sequences as an image

This function plots the collection of sequences as an image matrix.

viz_seqs_as_acgt_mat_from_seqs(
  rawSeqs,
  position_labels = NULL,
  xt_freq = 5,
  yt_freq = 100,
  col = c("darkgreen", "blue", "orange", "red"),
  savefilename = NULL,
  filetype = "PNG",
  fwidth = 450,
  fheight = 900,
  funits = "px"
)

Arguments

rawSeqs
The sequences as a DNAStringSet object.

position_labels
The labels to be used for the sequence positions. Default: Sequence positions are labeled from 1 to the length of the sequences.

xt_freq
The x-axis tick frequency.

yt_freq
The y-axis tick frequency.

col
A vector of four colors used for the DNA bases A, C, G, and T (in that order)

savefilename
Specify the filename (with extension) for saving the plot to disk

filetype
Specify the file type, namely PNG, JPEG, TIFF.

fwidth
Specify the width for the plot. This depends on the length of sequences.

fheight
Specify the height for the plot. This depends on the number of sequences.

funits
Specify the units in which the height and width are given.

Value

Nothing returned to the R interpreter.

See also