Plot ArchR Region Track

This function will plot the coverage at an input region in the style of a browser track. It allows for normalization of the signal which enables direct comparison across samples.

ArchRRegionTrack(ArchRProj = NULL, region = NULL,
  groupBy = "Clusters", useGroups = NULL, useCoverages = FALSE,
  plotSummary = c("bulkTrack", "featureTrack", "geneTrack"),
  sizes = c(10, 2, 4), features = getPeakSet(ArchRProj),
  geneSymbol = NULL, upstream = 50000, downstream = 50000,
  tileSize = 100, minCells = 25, normMethod = "ReadsInTSS",
  threads = 1, ylim = NULL, baseSize = 7, borderWidth = 0.4,
  tickWidth = 0.4, facetbaseSize = 7,
  geneAnno = getGeneAnnotation(ArchRProj), title = "", ...)

Arguments

ArchRProj
An ArchRProject object.

region
A GRanges region that indicates the region to be plotted. If more than one region exists in the GRanges object, all will be plotted. If no region is supplied, then the "geneSymbol" argument can be used to center the plot window at the transcription start site of the gene passed to "geneSymbol".

groupBy
A string that indicates how samples should be grouped. This string corresponds to one of the standard or user-supplied metadata columns (for example, "Clusters"). Samples with the same value annotated in this metadata column will be grouped together and the average signal will be used.

useGroups
A boolean value that indicates whether samples should be grouped together to produce average tracks. Only TRUE/FALSE are accepted.

useCoverages
QQQ use group coverages for track plotting

plotSummary
A character vector containing the features to be potted. Possible values include "bulkTrack" (the ATAC-seq signal), "featureTrack" (i.e. the peak bed regions), and "geneTrack" (line diagrams of genes with introns and exons shown. Blue-colored genes are on the minus strand and red-colored genes are on the plus strand).

sizes
A numeric vector containing 3 values that indicate the sizes of the individual components passed in plotSummary. The order must be the same as plotSummary.

features
A GRanges object containing the "features" to be plotted (This should be thought of as a bed track. i.e. the set of peaks obtained using getPeakSet(ArchRProj)).

geneSymbol
If "region" is not supplied, plotting can be centered at the transcription start site corresponding to the gene symbol(s) passed here.

upstream
The number of basepairs upstream of the transcription start site of "geneSymbol" to extend the plotting window. If "region" is supplied, this argument is ignored.

downstream
The number of basepairs downstream of the transcription start site of "geneSymbol" to extend the plotting window. If "region" is supplied, this argument is ignored.

tileSize
The numeric width of the tile/bin in basepairs for plotting ATAC-seq signal tracks. All insertions in a single bin will be summed.

normMethod
The name of the column in cellColData object by which normalization should be performed. The recommended and default value is "ReadsInTSS" which simultaneously normalizes tracks based on sequencing depth and sample data quality.

threads
The number of threads to use for parallel execution.

ylim
QQQ The numeric y-axis limit to be used for for bulkTrack plotting. If not provided, the y-axis limit will be QQQ.

baseSize
QQQ The numeric font size to be used in the plot. This applies to all plot labels.

borderWidth
The numeric line width to be used for plot borders.

tickWidth
The numeric line width to be used for axis tick marks.

geneAnno
QQQ The geneAnnotation objection to be used for plotting the "geneTrack" object. This must be a TxDB object.

title
verbose sections

...
additional args