Add Group Coverages to ArchR Project

This function will merge cells within each group into an insertion coverage file

addGroupCoverages(ArchRProj, groupBy = "Clusters", useLabels = TRUE,
  minCells = 40, maxCells = 500, maxFragments = 25 * 10^6,
  minReplicates = 2, maxReplicates = 5, sampleRatio = 0.8,
  kmerLength = 6, threads = 16, parallelParam = "mclapply",
  force = FALSE, verboseHeader = TRUE, verboseAll = FALSE, ...)

Arguments

ArchRProj
ArchRProject

groupBy
group cells by this column in cellColData

useLabels
use sample labels to create sample guided subgroupings as pseudo replicates

minCells
minimum cells per group for coverage files

maxCells
maximum cells per group for coverage files

maxFragments
maximum fragments per group for coverage files (this prevents large files created for optimizing memory)

minReplicates
minimum replicates for group for coverage files

maxReplicates
maximum replicates for group for coverage files

sampleRatio
sampling ratio for pseudo replicates when needed

kmerLength
kmer length for adding Tn5 bias estimation

threads
number of threads

parallelParam
parallel parameters for batch style execution

force
force creating coverage files if existed

verboseHeader
verbose sections

verboseAll
verbose sections and subsections

...
additional args