Add motif annotations to an ArchRProject

This function adds information about which peaks contain motifs to a given ArchRProject. For each peak, a binary value is stored indicating whether each motif is observed within the peak region.

addMotifAnnotations(
  ArchRProj = NULL,
  motifSet = "cisbp",
  name = "Motif",
  species = NULL,
  collection = "CORE",
  motifPWMs = NULL,
  cutOff = 5e-05,
  width = 7,
  version = 2,
  force = FALSE,
  logFile = createLogFile("addMotifAnnotations"),
  ...
)

Arguments

ArchRProj
An ArchRProject object.

motifSet
The motif set to be used for annotation. Options include: (i) "JASPAR2016", "JASPAR2018", "JASPAR2020" which gives the 2016, 2018 or 2020 version of JASPAR motifs or (ii) one of "cisbp", "encode", or "homer" which gives the corresponding motif sets from the chromVAR package.

name
The name of the peakAnnotation object to be stored in the provided ArchRProject

species
The name of the species relevant to the supplied ArchRProject. This is used for identifying which motif to be used from CisBP/JASPAR. By default, this function will attempt to guess the species based on the value from getGenome().

collection
If one of the JASPAR motif sets is used via motifSet, this parameter allows you to indicate the JASPAR collection to be used. See getMatrixSet() from TFBSTools for all options to supply for collection.

motifPWMs
A custom set of motif PWMs as a PWMList for adding motif annotations.

cutOff
The p-value cutoff to be used for motif search. The p-value is determined vs a background set of sequences (see MOODS for more details on this determination).

width
The width in basepairs to consider for motif matches. See the motimatchr package for more information.

version
An integer specifying version 1 or version 2 of chromVARmotifs see github for more info GreenleafLab/chromVARmotifs.

force
A boolean value indicating whether to force the peakAnnotation object indicated by name to be overwritten if it already exists in the given ArchRProject.

logFile
The path to a file to be used for logging ArchR output.

...
Additional parameters to be passed to TFBSTools::getMatrixSet for getting a PWM object.