Add Reproducible Peak Set to ArchR Project

This function will get insertions from coverage files call peaks and merge to get a Union Reproducible Peak Set

addReproduciblePeakSet(ArchRProj = NULL, groupBy = "Clusters",
  reproducibility = "2", peaksPerCell = 500, maxPeaks = 150000,
  minCells = 25, excludeChr = c("chrM", "chrY"),
  pathToMacs2 = findMacs2(), genomeSize = NULL, shift = -75,
  extsize = 150, method = "q", cutOff = 0.1, extendSummits = 250,
  promoterDist = 500, genomeAnno = getGenomeAnnotation(ArchRProj),
  geneAnno = getGeneAnnotation(ArchRProj),
  additionalParams = "--nomodel --nolambda", threads = 1,
  parallelParam = "mclapply", force = FALSE, verboseHeader = TRUE,
  verboseAll = FALSE, ...)

Arguments

ArchRProj
ArchRProject

groupBy
use groupings for peak calling matching group coverage files

reproducibility
reproducibility for peak calling (string that is a function of n)

peaksPerCell
number of peaks that can be identified per cell on average

excludeChr
exclude chromosomes from peak calling

pathToMacs2
path to macs2 executable (see Macs2)

genomeSize
genome size for peak calling (see Macs2)

shift
shift of Tn5 insertions (<- | ) (see Macs2)

extsize
extension of Tn5 insertions (|<- ->|) (see Macs2)

method
significance method for Macs2 (see Macs2)

cutOff
significance cutoff for Macs2 (see Macs2)

extendSummits
extend peak summits for final fixed-width peaks

promoterDist
promoter distance from TSS for annotating peaks

genomeAnno
genome annotation for ArchRProject

geneAnno
gene annotation for ArchRProject

additionalParams
additional parameters to pass to Macs2 (see Macs2)

threads
number of threads for parallel execution

parallelParam
parallel parameters for batch style execution

force
force creating peakSet if existed

verboseHeader
verbose sections

verboseAll
verbose sections and subsections

...
additional args