Add a Reproducible Peak Set to an ArchRProject

This function will get insertions from coverage files, call peaks, and merge peaks to get a "Union Reproducible Peak Set"

addReproduciblePeakSet(ArchRProj = NULL, groupBy = "Clusters",
  reproducibility = "2", peaksPerCell = 500, maxPeaks = 150000,
  minCells = 25, excludeChr = c("chrM", "chrY"),
  pathToMacs2 = findMacs2(), genomeSize = NULL, shift = -75,
  extsize = 150, method = "q", cutOff = 0.1, extendSummits = 250,
  promoterDist = 500, genomeAnno = getGenomeAnnotation(ArchRProj),
  geneAnno = getGeneAnnotation(ArchRProj),
  additionalParams = "--nomodel --nolambda", threads = 1,
  parallelParam = "mclapply", force = FALSE, verboseHeader = TRUE,
  verboseAll = FALSE, ...)

Arguments

ArchRProj
An ArchRProject object.

groupBy
The name of the column in cellColData to use for grouping cells together for peak calling.

reproducibility
A string that indicates how peak reproducibility should be handled. This string is dynamic and can be a function of n where n is the number of samples being assessed. For example, reproducibility = "2" means at least 2 samples must have a peak call at this locus and reproducibility = "(n+1)/2" means that the majority of samples must have a peak call at this locus.

peaksPerCell
The limit of number of peaks that can be identified per cell (this is useful for controlling how many peaks can be called from low cell groups).

maxPeaks
A numeric threshold for the maximum peaks to retain per group in groupBy in the union reproducible peak set.

minCells
The minimum number of unique cells that was used to create the coverage files on which peaks are called. This is important to allow for exclusion of pseudo-bulk replicates derived from very low cell numbers.

excludeChr
A character vector containing the seqnames of the chromosomes that should be excluded from peak calling.

pathToMacs2
The full path to the MACS2 executable.

genomeSize
The genome size to be used for MACS2 peak calling (see MACS2 documentation).

shift
The number of basepairs to shift each Tn5 insertion. When combined with extsize this allows you to create proper fragments, centered at the Tn5 insertion site, for use with MACS2 (see MACS2 documentation).

extsize
The number of basepairs to extend the MACS2 fragment after shift has been applied. When combined with extsize this allows you to create proper fragments, centered at the Tn5 insertion site, for use with MACS2 (see MACS2 documentation).

method
The method to use for significance testing in MACS2. Options are "p" for p-value and "q" for q-value. When combined with cutOff this gives the method and significance threshold for peak calling (see MACS2 documentation).

cutOff
The numeric significance cutoff for the testing method indicated by method (see MACS2 documentation).

extendSummits
The number of basepairs to extend peak summits (in both directions) to obtain final fixed-width peaks. For example, extendSummits = 250 will create 501-bp fixed-width peaks from the 1-bp summits.

promoterDist
The maximum distance in basepairs from the peak summit to the nearest transcriptional start site to allow for a peak to be annotated as a "promoter" peak.

genomeAnno
The genomeAnnotation (see createGenomeAnnotation) is used for downstream analyses for genome information such as nucleotide information (GC info) or chromosome sizes.

geneAnno
The geneAnnotation (see createGeneAnnotation) is used for peak labeling as promoter etc.

additionalParams
A string of additional parameters to pass to MACS2 (see MACS2 documentation).

threads
The number of threads to be used for parallel computing.

parallelParam
A list of parameters to be passed for biocparallel/batchtools parallel computing.

force
A boolean value indicating whether to force the reproducible peak set to be overwritten if it already exist in the given ArchRProject peakSet.

verboseHeader
A boolean value that determines whether standard output includes verbose sections.

verboseAll
A boolean value that determines whether standard output includes verbose subsections.

...
additional args