archR: A package for de novo discovery of different sequence architectures composed of combination(s) of sequence elements (motifs).

Given a set of DNA sequences, archR enables discovery of clusters of sequence with characteristic sequence architectures.

Call this function to process a data set using archR.

archR(
  config,
  seqsMat,
  seqsRaw,
  seqsPositions = NULL,
  thresholdItr = 3,
  setParsimony = c(FALSE, FALSE, FALSE),
  setOCollation = c(TRUE, FALSE, FALSE),
  fresh = TRUE,
  UseOC = NULL,
  oDir = NULL
)

Arguments

config
archR configuration object as returned by archRSetConfig. This is a required argument.

seqsMat
A matrix of one-hot encoded sequences with sequences along columns. This is a required argument.

seqsRaw
A Biostrings::DNAStringSet object. The FASTA sequences as a DNAStringSet object. This argument required argument.

seqsPositions
Vector. Specify the tick labels for sequence positions. Default is NULL.

thresholdItr
Numeric. Specify the number of iterations to perform. Default is 3.

setParsimony
Logical vector. Specify if model selection by cross-validation should prefer parsimonious solutions. Not required when stability-based model selection is chosen. Length of the vector should match number of threshold iterations specified in thresholdItr argument. TRUE denotes parsimony is followed, FALSE otherwise.

setOCollation
Logical vector. Specify for every iteration of archR if collation of clusters from outer chunks should be performed. TRUE denotes clusters are collated, FALSE otherwise.

fresh
Logical. Specify if this is (not) a fresh run. Because archR enables checkpointing, it is possible to perform additional iterations upon clusters from an existing archR result (or a checkpoint) object. See UseOC argument. For example, when processing a set of FASTA sequences, if an earlier call to archR performed two iterations, and now you wish to perform a third, the arguments `fresh` and `UseOC` can be used. Simply set `fresh` to FALSE and assign the sequence clusters from iteration two from the earlier result to `UseOC`. As of v0.1.3, with this setting, archR returns a new result object as if the additional iteration performed is the only iteration.

UseOC
List. Clusters to be further processed with archR. These can be from a previous archR result (in which case use archRresult$clustSol$clusters), or simply clusters from any other method. Warning: This has not been rigorously tested yet (v0.1.3).

oDir
Character. Specify the output directory with its path. archR will create this directory. If a directory with the given name exists at the given location, archR will add a suffix to the directory name. This change is reported to the user. Default is NULL. When NULL, just the result is returned, and no plots or checkpoints or result is written to disk.

Value

A nested list of elements as follows:

seqsClustLabels

A list with cluster labels for all sequences per iteration of archR. The cluster labels as stored as characters.

clustBasisVectors

A list with information on NMF basis vectors per iteration of archR. Per iteration, there are two variables `nBasisVectors` storing the number of basis vectors after model selection, and `basisVectors`, a matrix storing the basis vectors themselves. Dimensions of the `basisVectors` matrix are 4*L x nBasisVectors (mononucleotide case) or 16*L x nBasisVectors (dinucleotide case).

clustSol

The clustering solution obtained upon processing the raw clusters from the last iteration of archR's result. This is handled internally by the function reorder_archRresult using the default setting of Euclidean distance and ward.D linkage hierarchical clustering.

rawSeqs

The input sequences as a DNAStringSet object.

timeInfo

Stores the time taken (in minutes) for processing each iteration. This element is added only if `timeFlag` is set to TRUE in config.

config

The configuration used for processing.

call

The function call itself.

Details

The archR package provides three categories of important functions: related to data preparation and manipulation, performing non-negative matrix factorization, performing clustering, and visualization-related functions.

Functions for data preparation and manipulation

prepare_data_from_FASTA

get_one_hot_encoded_seqs

Functions for visualizations

plot_ggseqlogo

plot_ggheatmap

viz_basis_vectors_as_heatmap

viz_basis_vectors_as_seqlogo

viz_basis_vectors_in_combined_heatmaps_seqlogos

plot_arch_for_clusters

viz_seqs_as_acgt_mat_from_seqs

Examples

inputFastaFilename <- system.file("extdata", "example_data.fa", 
                        package = "archR", mustWork = TRUE)

# Specifying 'dinuc' generates dinucleotide features
inputSeqsMat <- archR::prepare_data_from_FASTA(inputFastaFilename,
    sinuc_or_dinuc = "dinuc")

#> Sequences OK, 

#> Read 200 sequences

#> Generating dinucleotide profiles

inputSeqsRaw <- archR::prepare_data_from_FASTA(inputFastaFilename, 
    rawSeq = TRUE)

# Set archR configuration
archRconfig <- archR::archRSetConfig(
    parallelize = TRUE,
    nCoresUse = 2,
    nIterationsUse = 100,
    kMin = 1,
    kMax = 20,
    modSelType = "stability",
    tol = 10^-4,
    bound = 10^-8,
    innerChunkSize = 100,
    flags = list(debugFlag = TRUE, timeFlag = TRUE, verboseFlag = TRUE,
        plotVerboseFlag = FALSE)
)

# Run archR 
archRresult <- archR::archR(config = archRconfig,
                          seqsMat = inputSeqsMat,
                          seqsRaw = inputSeqsRaw,
                          seqsPositions = seq(1,100,by=1),
                          thresholdItr = 2)

#> Asked for parallelization

#> Making cluster here

#> Parallelization with 2 cores

#> list(con = 5, host = "localhost", rank = 1)list(con = 6, host = "localhost", rank = 2)

#> Model selection by factor stability

#> Tolerance: 1e-04 & Bound: 1e-08

#> === Iteration 1, 1 chunk(s) ===

#> [Outer chunk 1/1] [Size: 200]

#> Decision: Processing

#> [Inner chunk 1/2] [Size: 100]

#> Performing stability-based model selection

#> Tolerance: 1e-04 & Bound : 1e-08

#> Checking K = 1

#> AmariTypeDist : 2.15315980533364e-17

#> Checking K = 2

#> AmariTypeDist : 7.48895894792606e-17

#> Change is : 0

#> Checking K = 3

#> AmariTypeDist : 1.82659723485804e-16

#> Change is : 1

#> Checking K = 4

#> AmariTypeDist : 4.8923605148522e-07

#> This amariType Distance = 4.8923605148522e-07

#> Best K for this subset: 3

#> Fetching 3 clusters

#> Best Q2 giving run found: 0.0416069700312701

#> Fetching 3 cluster(s) w/ NMF scores

#> [Inner chunk 2/2] [Size: 100]

#> Performing stability-based model selection

#> Tolerance: 1e-04 & Bound : 1e-08

#> Checking K = 1

#> AmariTypeDist : 2.51201977288924e-17

#> Checking K = 2

#> AmariTypeDist : 7.13234185516767e-18

#> Change is : 1

#> Checking K = 3

#> AmariTypeDist : 2.73885515838114e-14

#> Change is : 4

#> magnitude change higher than tolerance, 4

#> This amariType Distance = 2.73885515838114e-14

#> Would have choosen bestK as : 2

#> Checking K = 4

#> AmariTypeDist : 1.32872604209591e-08

#> This amariType Distance = 1.32872604209591e-08

#> Best K for this subset: 3

#> Fetching 3 clusters

#> Best Q2 giving run found: 0.0496099954482235

#> Fetching 3 cluster(s) w/ NMF scores

#> [Outer chunk 1/1] complete

#> [Iteration 1] complete

#> [Outer chunk 1/1] complete

#> current total factors: 6

#> Current total chunks for next iteration: 6

#> Managing clusters from outer chunk(s)

#> Collation this iter:TRUE; #InnerChunks: 2; #OuterChunks: 1

#> Decision for outer chunk collation: Yes

#> 

#> Hierarchical clustering

#> RE-ORDERING CLUSTERS

#> Cluster order by ward.D linkage w/ euclid distance:

#> New order: 2, 5, 1, 4, 3, 6

#> 2, 5, 2.94351671321916

#> 5, 1, 5.93222437689053

#> 1, 4, 2.4270593689496

#> 4, 3, 5.53264304797784

#> 3, 6, 3.00071774134856

#> New order: 2, 5, 1, 4, 3, 6

#> === CALLING get_clusters FUNCTION === 

#> Max. silhouette score at index = 13, height:3.0270593689496

#> Final #Clusters using silhouette values: 3

#> c(1, 4)
#> c(2, 5)
#> c(3, 6)

#> Final #Clusters: 3

#> c(1, 4)
#> c(2, 5)
#> c(3, 6)

#> #Clusters: 3

#> === DONE ===

#> ClustList: c(1, 4)
#> c(2, 5)
#> c(3, 6)

#> === This is what I got ===

#> c(3, 5, 7, 8, 10, 12, 13, 16, 17, 18, 19, 22, 25, 30, 31, 36, 37, 38, 39, 41, 43, 45, 47, 49, 51, 53, 54, 55, 56, 58, 59, 60, 61, 64, 67, 69, 70, 73, 77, 84, 89, 91, 94, 95, 98)
#> c(1, 2, 6, 11, 20, 23, 32, 46, 50, 52, 63, 68, 71, 72, 75, 80, 82, 83, 85, 86, 87, 88, 92, 99)
#> c(4, 9, 14, 15, 21, 24, 26, 27, 28, 29, 33, 34, 35, 40, 42, 44, 48, 57, 62, 65, 66, 74, 76, 78, 79, 81, 90, 93, 96, 97, 100)
#> c(103, 106, 107, 109, 115, 117, 118, 119, 120, 121, 122, 124, 125, 126, 128, 130, 132, 133, 134, 135, 136, 140, 141, 142, 144, 150, 152, 153, 154, 156, 159, 160, 161, 162, 164, 166, 170, 173, 175, 177, 180, 183, 184, 185, 186, 188, 189, 191, 193, 194, 195, 196, 198, 199, 200)
#> c(102, 104, 105, 112, 114, 116, 123, 131, 137, 138, 143, 145, 146, 147, 148, 151, 158, 165, 167, 168, 171, 176, 179, 182, 190, 197)
#> c(101, 108, 110, 111, 113, 127, 129, 139, 149, 155, 157, 163, 169, 172, 174, 178, 181, 187, 192)

#> === INFO ===

#> 1, 4

#> Size:100

#> === INFO ===

#> 2, 5

#> Size:50

#> === INFO ===

#> 3, 6

#> Size:50

#> === I am returning ===

#> c(3, 5, 7, 8, 10, 12, 13, 16, 17, 18, 19, 22, 25, 30, 31, 36, 37, 38, 39, 41, 43, 45, 47, 49, 51, 53, 54, 55, 56, 58, 59, 60, 61, 64, 67, 69, 70, 73, 77, 84, 89, 91, 94, 95, 98, 103, 106, 107, 109, 115, 117, 118, 119, 120, 121, 122, 124, 125, 126, 128, 130, 132, 133, 134, 135, 136, 140, 141, 142, 144, 150, 152, 153, 154, 156, 159, 160, 161, 162, 164, 166, 170, 173, 175, 177, 180, 183, 184, 185, 186, 188, 189, 191, 193, 194, 195, 196, 198, 199, 200)
#> c(1, 2, 6, 11, 20, 23, 32, 46, 50, 52, 63, 68, 71, 72, 75, 80, 82, 83, 85, 86, 87, 88, 92, 99, 102, 104, 105, 112, 114, 116, 123, 131, 137, 138, 143, 145, 146, 147, 148, 151, 158, 165, 167, 168, 171, 176, 179, 182, 190, 197)
#> c(4, 9, 14, 15, 21, 24, 26, 27, 28, 29, 33, 34, 35, 40, 42, 44, 48, 57, 62, 65, 66, 74, 76, 78, 79, 81, 90, 93, 96, 97, 100, 101, 108, 110, 111, 113, 127, 129, 139, 149, 155, 157, 163, 169, 172, 174, 178, 181, 187, 192)

#> Updating sequence cluster labels

#> #Clusters: 3

#> Iteration 1 completed in 1.27 mins

#> 1.27 mins ellapsed since start

#> === Iteration 2, 3 chunk(s) ===

#> [Outer chunk 1/3] [Size: 100]

#> Decision: Processing

#> [Inner chunk 1/1] [Size: 100]

#> Performing stability-based model selection

#> Tolerance: 1e-04 & Bound : 1e-08

#> Checking K = 1

#> AmariTypeDist : 1.61711272879745e-17

#> Checking K = 2

#> AmariTypeDist : 4.75160915776556e-11

#> Change is : 6

#> magnitude change higher than tolerance, 4

#> This amariType Distance = 4.75160915776556e-11

#> Would have choosen bestK as : 1

#> Checking K = 3

#> AmariTypeDist : 0.0035786011862791

#> This amariType Distance = 0.0035786011862791

#> Best K for this subset: 2

#> Fetching 2 clusters

#> Best Q2 giving run found: 0.114541701362898

#> Fetching 2 cluster(s) w/ NMF scores

#> [Outer chunk 1/3] complete

#> [Outer chunk 1/3] complete

#> current total factors: 2

#> Current total chunks for next iteration: 2

#> [Outer chunk 2/3] [Size: 50]

#> Decision: Processing

#> [Inner chunk 1/1] [Size: 50]

#> Performing stability-based model selection

#> Tolerance: 1e-04 & Bound : 1e-08

#> Checking K = 1

#> AmariTypeDist : 1.59019823123078e-17

#> Checking K = 2

#> AmariTypeDist : 2.57849364933938e-06

#> This amariType Distance = 2.57849364933938e-06

#> Best K for this subset: 1

#> Fetching 1 clusters

#> Best Q2 giving run found: 0.197789465153941

#> Fetching 1 cluster(s) w/ NMF scores

#> [Outer chunk 2/3] complete

#> [Outer chunk 2/3] complete

#> current total factors: 3

#> Current total chunks for next iteration: 3

#> [Outer chunk 3/3] [Size: 50]

#> Decision: Processing

#> [Inner chunk 1/1] [Size: 50]

#> Performing stability-based model selection

#> Tolerance: 1e-04 & Bound : 1e-08

#> Checking K = 1

#> AmariTypeDist : 3.17142496327267e-17

#> Checking K = 2

#> AmariTypeDist : 7.67125992969052e-05

#> This amariType Distance = 7.67125992969052e-05

#> Best K for this subset: 1

#> Fetching 1 clusters

#> Best Q2 giving run found: 0.153271954789996

#> Fetching 1 cluster(s) w/ NMF scores

#> [Outer chunk 3/3] complete

#> [Iteration 2] complete

#> [Outer chunk 3/3] complete

#> current total factors: 4

#> Current total chunks for next iteration: 4

#> Managing clusters from outer chunk(s)

#> Collation this iter:FALSE; #InnerChunks: 1; #OuterChunks: 3

#> Decision for outer chunk collation: No

#> Updating sequence cluster labels

#> #Clusters: 4

#> Iteration 2 completed in 0.814 mins

#> 2.08 mins ellapsed since start

#> Hierarchical clustering

#> RE-ORDERING CLUSTERS

#> Cluster order by ward.D linkage w/ euclid distance:

#> New order: 3, 4, 1, 2

#> 3, 4, 6.00792293609081

#> 4, 1, 5.669887508159

#> 1, 2, 3.67152457274261

#> New order: 3, 4, 1, 2

#> === CALLING get_clusters FUNCTION === 

#> No element pairs close enough by given distance threshold: 3

#> #Clusters: 4

#> === DONE ===

#> ClustList: 3
#> 4
#> 1
#> 2

#> Performed factor clustering, returning object:

#> Returning: 3
#> 4
#> 1
#> 2

#> Updating sequence cluster labels

#> #Clusters: 4

#> Stopping cluster...

#> archR exiting, 2.08 mins