Create Arrow Files

This function will create an Arrow Files from input files for downstream analysis

createArrowFiles(inputFiles = NULL, sampleNames = NULL,
  outputNames = paste0("./", sampleNames), geneAnno = NULL,
  genomeAnno = NULL, filterFrags = 1000, filterTSS = 4,
  removeFilteredCells = TRUE, minFrags = 500,
  outDir = "QualityControl", nucLength = 147, TSSParams = list(),
  excludeChr = c("chrM", "chrY"), nChunk = 5, bcTag = "qname",
  gsubExpression = NULL, bamFlag = NULL, offsetPlus = 4,
  offsetMinus = -5, addTileMat = TRUE, TileMatParams = list(),
  addGeneScoreMat = TRUE, GeneScoreMatParams = list(), force = FALSE,
  threads = 1, parallelParam = NULL, verboseHeader = TRUE,
  verboseAll = FALSE, ...)

Arguments

inputFiles
input files (tabixFile, bamFile or textFile)

sampleNames
sample names corresponding to input files

outputNames
output names prefix (ie PBMC -> PBMC.arrow)

geneAnno
geneAnnotation input for TSS Scores etc.

genomeAnno
genomeAnnotation input for ChromSizes Nucleotide Information etc.

filterFrags
min fragments per cell to be filtered for analyses such as tileMat etc.

filterTSS
min TSS Score per cell to be filtered for analyses such as tileMat etc.

removeFilteredCells
remove fragments corresponding to cells pass filterFrags and filterTSS

minFrags
min fragments per cell to be immediately filtered

outDir
out directory for QC information from sample to be plotted / saved

nucLength
nucleosome length for id'ing fragments as sub-, mono-, or multi-nucleosome spanning

TSSParams
TSS parameters for computing TSS scores

excludeChr
exclude these chromosomes from analysis downstream (does not apply to fragments)

nChunk
number of chunks per chromosome when reading in input files

bcTag
barcode tag location in bam file (see ScanBam in Rsamtools)

bamFlag
list of bam flags for reading in fragments from input files (see ScanBam in Rsamtools)

offsetPlus
Tn5 offset of "+" stranded insertion (see Buenrostro 2013)

offsetMinus
Tn5 offset of "-" stranded insertion (see Buenrostro 2013)

addTileMat
addTileMatrix to ArrowFiles

TileMatParams
additional parameters to pass to addTileMatrix (see addTileMatrix)

addGeneScoreMat
addGeneScoreMatrix to ArrowFiles

GeneScoreMatParams
additional parameters to pass to addGeneScoreMatrix (see addGeneScoreMatrix)

force
force creation of arrow files if already exist

threads
number threads for parallel execution

parallelParam
parallel parameters for batch style execution

...
additional args