A ggplot-based one-to-one dot plot wrapper function

ggOneToOne(x = NULL, y = NULL, size = 2, alpha = 1, xlabel = "x",
  ylabel = "y", title = "Correlation", min = 0.05, max = 0.9999,
  nPlot = 100 * 10^3, nKernel = 100, densityMax = 0.95,
  extend = 0.05, baseSize = 6, rastr = TRUE,
  pal = paletteContinuous(set = "blue_yellow"), ...)

Arguments

x
A numeric vector containing the x-axis values for each point.

y
A numeric vector containing the y-axis values for each point.

size
The numeric size of the points to plot.

alpha
A number indicating the transparency to use for each point. See `ggplot2` for more details.

xlabel
The label to plot for the x-axis.

ylabel
The label to plot for the y-axis.

title
The title of the plot.

min
QQQ xmin quantile [0,1]

max
QQQ xmax quantile [0,1]

nPlot
The number of points to plot. When this value is less than the total points, sample is used to extract random data points to plot.

nKernel
The value of n to use the kde2d from the MASS package.

densityMax
QQQ The quantile that should be represented by the maximum color on the continuous scale designated by `pal`. Above this value will be thresholded to the maximum color.

extend
QQQ A numeric value indicating the fraction to extend the x-axis and y-axis beyond the maximum and minimum values if `xlim` and `ylim` are not provided. For example, 0.05 will extend the x-axis and y-axis by 5

baseSizeQQQ The base font size to use in the plot.

rastrA boolean value that indicates whether the plot should be rasterized. This does not rasterize lines and labels, just the internal portions of the plot.

palQQQ The name or numeric index of a custom palette from `ArchRPalettes` to use for the density in the one-to-one plot. In most cases, a continuous palette should be used.