Plot a Heatmap of Identified Marker Features

This function will plot a heatmap of the results from markerFeatures

This function will plot a heatmap of the results from markerFeatures

markerHeatmap(seMarker, cutOff = "FDR <= 0.001 & Log2FC >= 0.1",
  log2Norm = TRUE, scaleTo = 10^4, scaleRows = TRUE, limits = c(-2,
  2), grepExclude = NULL, pal = NULL, binaryClusterRows = TRUE,
  labelMarkers = NULL, labelTop = NULL, labelRows = FALSE,
  returnMat = FALSE, ...)

markerHeatmap(seMarker, cutOff = "FDR <= 0.001 & Log2FC >= 0.1",
  log2Norm = TRUE, scaleTo = 10^4, scaleRows = TRUE, limits = c(-2,
  2), grepExclude = NULL, pal = NULL, binaryClusterRows = TRUE,
  labelMarkers = NULL, labelTop = NULL, labelRows = FALSE,
  returnMat = FALSE, ...)

Arguments

seMarker
Summarized Experiment result from markerFeatures

log2Norm
log2 Normalization prior to plotting set true for counting assays (not DeviationsMatrix!)

scaleTo
scale to prior to log2 Normalization, if log2Norm is FALSE this does nothing

scaleRows
compute row z-scores on matrix

limits
heatmap color limits

grepExclude
remove features by grep

pal
palette for heatmap, default will use solar_extra

binaryClusterRows
fast clustering implementation for row clustering by binary sorting

labelMarkers
label specific markers by name on heatmap (matches rownames of seMarker)

labelTop
label the top features for each column in seMarker

labelRows
label all rows

returnMat
return final matrix that is used for plotting heatmap

...
additional args

cutoff
Logical Statement for Cutoff to Be called a Marker a statement containing assayNames from seMarker

seMarker
Summarized Experiment result from markerFeatures

cutoff
Logical Statement for Cutoff to Be called a Marker a statement containing assayNames from seMarker

log2Norm
log2 Normalization prior to plotting set true for counting assays (not DeviationsMatrix!)

scaleTo
scale to prior to log2 Normalization, if log2Norm is FALSE this does nothing

scaleRows
compute row z-scores on matrix

limits
heatmap color limits

grepExclude
remove features by grep

pal
palette for heatmap, default will use solar_extra

binaryClusterRows
fast clustering implementation for row clustering by binary sorting

labelMarkers
label specific markers by name on heatmap (matches rownames of seMarker)

labelTop
label the top features for each column in seMarker

labelRows
label all rows

returnMat
return final matrix that is used for plotting heatmap

...
additional args