Visualize an Embedding from ArchR Project

This function will plot an embedding stored in an ArchRProject

plotEmbedding(ArchRProj = NULL, embedding = "UMAP",
  colorBy = "cellColData", name = "Sample", log2Norm = NULL,
  imputeWeights = NULL, pal = NULL, size = 0.5, rastr = TRUE,
  quantCut = c(0.01, 0.99), discreteSet = NULL, continuousSet = NULL,
  randomize = TRUE, keepAxis = FALSE, baseSize = 6, plotAs = NULL,
  plotParams = list(), ...)

Arguments

ArchRProj
An ArchRProject object.

embedding
The name of the embedding stored in the ArchRProject to be plotted. See computeEmbedding() for more information.

colorBy
A string indicating whether points in the plot should be colored by a column in cellColData ("cellColData") or by a data matrix in the corresponding ArrowFiles (i.e. "GeneScoreMatrix", "MotifMatrix", "PeakMatrix").

name
The name of the column in cellColData or the featureName/rowname of the data matrix to be used for plotting. For example if colorBy is cellColData then name refers to a column name in the cellcoldata (see getCellcoldata()), if colorBy is GeneScoreMatrix then name refers to a gene name which can be listed by getFeatures(ArchRProj, useMatrix = "GeneScoreMatrix").

log2Norm
A boolean value indicating whether a log2 transformation should be performed on the values (if continuous) in plotting.

imputeWeights
The weights to be used for imputing numerical values for each cell as a linear combination of other cells values. See addImputationWeights() and getImutationWeights() for more information.

pal
A custom palette (see paletteDiscrete or ArchRPalettes) used to override discreteSet/continuousSet for coloring vector.

size
A number indicating the size of the points to plot if plotAs is set to "points".

rastr
A boolean value that indicates whether the plot should be rasterized. This does not rasterize lines and labels, just the internal portions of the plot.

quantCut
If this is not NULL, a quantile cut is performed to threshold the top and bottom of the distribution of numerical values. This prevents skewed color scales caused by strong outliers. The format of this should be c(x,y) where x is the lower threshold and y is the upper threshold. For example, quantileCut = c(0.025,0.975) will take the 2.5th percentile and 97.5 percentile of values and set values below/above to the value of the 2.5th and 97.5th percentile values respectively.

discreteSet
The name of a discrete palette from ArchRPalettes for visualizing colorBy in the embedding if a discrete color set is desired.

continuousSet
The name of a continuous palette from ArchRPalettes for visualizing colorBy in the embedding if a continuous color set is desired.

randomize
A boolean value that indicates whether to randomize points prior to plotting to prevent cells from one cluster being present at the front of the plot.

keepAxis
A boolean value that indicates whether the x and y axis ticks and labels should be plotted.

baseSize
The base font size to use in the plot.

plotAs
A string that indicates whether points ("points") should be plotted or a hexplot ("hex") should be plotted. By default if colorBy is numeric this is "hex".

plotParams
Additional parameters to pass to ggPoint() or ggHex().

...
additional args