Plot footprints for an ArchRProject

This function will plot footprints for all samples in a given ArchRProject or a properly-formatted Summarized Experiment

plotFootprints(input = NULL, positions = NULL,
  plotName = "Plot-Footprints", groupBy = "Clusters",
  useGroups = NULL, pal = NULL, flank = 250, flankNorm = 50,
  smoothWindow = 10, minCells = 25, nTop = NULL,
  normMethod = "none", height = 6, width = 4, addDOC = TRUE,
  useSink = TRUE, threads = 1, verboseHeader = TRUE,
  verboseAll = FALSE, ...)

Arguments

input
An ArchRProject object or Footprint Summarized Experiment

positions
QQQ A GenomicRangesList, a list, or a SimpleList object containing the positions to incorporate into the footprint. Each position should be QQQ.

groupBy
QQQ The column name in sampleColData to use for grouping multiple samples together prior to footprinting.

useGroups
QQQ

pal
The name or numeric index of a custom palette from ArchR_palettes to use for plotting the lines corresponding to the footprints.

flank
QQQ The number of basepairs from the position center to consider as the flank.

flankNorm
QQQ The number of basepairs to consider at the edge of the flank region to be used for footprint normalization.

smoothWindow
QQQ The size in basepairs of the sliding window to be used for smoothing of the footprint signal.

nTop
QQQ The number of positions to consider. Only the top nTop positions based on QQQ will be considered for the footprint.

normMethod
QQQ The name of the normalization method to use to normalize the footprint relative to the Tn5 insertion bias. Options include QQQ.

threads
The number threads to be used for parallel computing.

verboseHeader
A boolean value that determines whether standard output includes verbose sections.

verboseAll
A boolean value that determines whether standard output includes verbose subsections.

...
additional args