QQQ Visualize a Trajectory from ArchR Project

plotTrajectory(ArchRProj = NULL, embedding = "UMAP",
  trajectory = "Trajectory", colorBy = "colData",
  name = "Trajectory", log2Norm = NULL, pal = NULL, size = 0.5,
  rastr = TRUE, quantCut = c(0.05, 0.95), quantHex = 0.5,
  discreteSet = NULL, continuousSet = NULL, randomize = TRUE,
  keepAxis = FALSE, baseSize = 6, addArrow = TRUE, plotAs = NULL,
  plotParams = list(), ...)

Arguments

ArchRProj
An ArchRProject object.

embedding
The name of the embedding to use to visualize the given trajectory. See ArchR::computeEmbedding() for more information.

trajectory
The name of the trajectory as a column in cellColData to plot.

colorBy
QQQ UNCLEAR A string indicating whether points/hexes in the plot should be colored by cellColData ("cellColData") or by the gene activity score matrix ("GeneScoreMatrix").

name
The name of the trajectory as a column in cellColData to plot.

log2Norm
QQQ A boolean value indicating whether a log2 transformation should be performed on the values in QQQ plotting.

pal
QQQ The name or numeric index of a custom palette from ArchRPalettes to use for QQQ.

size
QQQ A number indicating the size of the points to plot if plotAs is set to "points".

rastr
A boolean value that indicates whether the plot should be rasterized. This does not rasterize lines and labels, just the internal portions of the plot.

quantCut
If this is not null, a quantile cut is performed to threshold the top and bottom of the distribution. This prevents skewed color scales caused by strong outliers. The format of this should be c(x,y) where x is the upper threshold and y is the lower threshold. For example, quantileCut = c(0.975,0.025) will take the top and bottom 2.5

quantHexQQQ UNCLEAR quantile evaluation for each hex in geom_hex

discreteSetQQQ The name or numeric index of a discrete palette from ArchRPalettes for visualizing QQQ in the embedding.

continuousSetQQQ The name or numeric index of a continuous palette from ArchRPalettes for visualizing QQQ in the embedding.

randomizeA boolean value that indicates whether to randomize points prior to plotting to prevent cells from one cluster being present at the front of the plot.

keepAxisA boolean value that indicates whether the x and y axis ticks and labels should be plotted.

baseSizeQQQ The base font size to use in the plot.

addArrowQQQ A boolean value that indicates whether QQQ.

plotAsA string that indicates whether points ("points") should be plotted or a hexplot ("hex") should be plotted.

plotParamsAdditional parameters to pass to ggPoint() or ggHex().

...additional args